+++ lib/src/io/http.dart
@@ -4,7 +4,7 @@
 part of rikulo_io;
 
 /**
- * The request wrapper.
+ * The HTTP request wrapper.
  */
 class HttpRequestWrapper implements HttpRequest {
   ///The original HTTP request
@@ -45,7 +45,7 @@ class HttpRequestWrapper implements HttpRequest {
 }
 
 /**
- * The response wrapper.
+ * The HTTP response wrapper.
  */
 class HttpResponseWrapper implements HttpResponse {
   ///The original HTTP response
@@ -92,3 +92,78 @@ class HttpResponseWrapper implements HttpResponse {
   @override
   HttpConnectionInfo get connectionInfo => origin.connectionInfo;
 }
+
+/**
+ * The HTTP headers wrapper.
+ */
+class HttpHeadersWrapper extends HttpHeaders {
+  ///The original HTTP headers
+  final HttpHeaders origin;
+
+  HttpHeadersWrapper(HttpHeaders this.origin);
+
+  //@override
+  List<String> operator[](String name) => origin[name];
+  //@override
+  String value(String name) => origin.value(name);
+  //@override
+  void add(String name, Object value) {
+    origin.add(name, value);
+  }
+  //@override
+  void set(String name, Object value) {
+    origin.set(name, value);
+  }
+  //@override
+  void remove(String name, Object value) {
+    origin.remove(name, value);
+  }
+  //@override
+  void removeAll(String name) {
+    origin.removeAll(name);
+  }
+  //@override
+  void forEach(void f(String name, List<String> values)) {
+    origin.forEach(f);
+  }
+  //@override
+  void noFolding(String name) {
+    origin.noFolding(name);
+  }
+  @override
+  Date get date => origin.date;
+  @override
+  void set date(Date date) {
+    origin.date = date;
+  }
+  @override
+  Date get expires => origin.expires;
+  @override
+  void set expires(Date expires) {
+    origin.expires = expires;
+  }
+  @override
+  Date get ifModifiedSince => origin.ifModifiedSince;
+  @override
+  void set ifModifiedSince(Date ifModifiedSince) {
+    origin.ifModifiedSince = ifModifiedSince;
+  }
+  @override
+  String get host => origin.host;
+  @override
+  void set host(String host) {
+    origin.host = host;
+  }
+  @override
+  int get port => origin.port;
+  @override
+  void set port(int port) {
+    origin.port = port;
+  }
+  @override
+  ContentType get contentType => origin.contentType;
+  @override
+  void set contentType(ContentType contentType) {
+    origin.contentType = contentType;
+  }
+}