+++ CHANGES.md
@@ -2,9 +2,9 @@
 
 ##0.6.2
 
-January 9, 2013
+January 23, 2013
 
 **Features**
 
 * StringUtil, MapUtil and many others are added.
-* io.dart, supporting HttpRequestWrapper and others, is added
+* HttpRequestWrapper, HttpRespnseWrapper and HttpHeadersWrapper are added.